+++ crates/ui/src/root.rs
@@ -40,7 +40,7 @@ impl Root {
         return root;
     }
     // Add an element to the tree
-    pub fn add_element(&mut self, mut element: Element) -> u16 {
+    pub fn add_element(&mut self, mut element: Element) -> usize {
         // Get the ID of the element
         let element_id = self.smart_element_list.get_next_valid_id();
         element.id = element_id;
@@ -54,12 +54,12 @@ impl Root {
     // Remove an element from the three, and recursively remove it's children
     pub fn remove_element(&mut self, element: Element) {
         // Get all the children from this element, recursively
-        let mut output_elem_indices: Vec<u16> = Vec::new();
-        let mut elems_to_evaluate: Vec<u16> = Vec::new();
+        let mut output_elem_indices: Vec<usize> = Vec::new();
+        let mut elems_to_evaluate: Vec<usize> = Vec::new();
         elems_to_evaluate.extend(element.children);
         while elems_to_evaluate.len() > 0 {
             // We need to get the children of this element
-            let elem = self.smart_element_list.get_element(elems_to_evaluate[0] as u16).unwrap();
+            let elem = self.smart_element_list.get_element(elems_to_evaluate[0]).unwrap();
             let children = elem.children.clone();
             elems_to_evaluate.extend(children);
             elems_to_evaluate.remove(0);
@@ -68,15 +68,15 @@ impl Root {
 
     // ---- Actual root UI stuff ---- \\
     // Get an element from the root using it's id
-    pub fn get_element(&self, id: u16) -> &Element {
+    pub fn get_element(&self, id: usize) -> &Element {
         self.smart_element_list.get_element(id).unwrap()
     }
     // Get an element from the root using it's id
-    pub fn get_element_mut(&mut self, id: u16) -> &mut Element {
+    pub fn get_element_mut(&mut self, id: usize) -> &mut Element {
         self.smart_element_list.get_element_mut(id).unwrap()
     }
     // Get the state of a specific button element
-    pub fn get_button_state(&self, element_id: u16) -> &ButtonState {
+    pub fn get_button_state(&self, element_id: usize) -> &ButtonState {
         // Get the element
         let elem = self.smart_element_list.get_element(element_id).unwrap();
         let state = match elem.element_type {
@@ -86,7 +86,7 @@ impl Root {
         return state;
     }
     // Set the text of a text element
-    pub fn set_text_state(&mut self, element_id: u16, text: &str) {
+    pub fn set_text_state(&mut self, element_id: usize, text: &str) {
         // Get the element mutably
         let elem = self.smart_element_list.get_element_mut(element_id).unwrap();
         match elem.element_type {
